requirements.txt
fastapi
uvicorn
python-multipart
cohere
requests
python-dotenv
pinecone-client
PyMuPDF
